Raft take snapshot process is not proper
Here, the method is taking a snapshot "Per service", but somehow looping over the entire session list. This means at t will create a problem the time of restore, It will create all the sessions for the first service getting loaded.
sample effect :  If there are changes between the take-snapshot and reboot node, All will go missing from the Service.